+++ src/main/java/de/pt400c/defaultsettings/gui/Segment.java
@@ -14,8 +14,8 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 public abstract class Segment {
 	
 	protected final GuiScreen gui;
-	protected double posX;
-	protected double posY;
+	protected float posX;
+	protected float posY;
 	protected float width;
 	protected float height;
 	protected final boolean isPopupSegment;
@@ -29,23 +29,23 @@ public abstract class Segment {
 		this.isPopupSegment = popupSegment;
 	}
 	
-	public abstract void render(float mouseX, float mouseY, float partialTicks);
+	public abstract void render(int mouseX, int mouseY, float partialTicks);
 	
-	public void customRender(float mouseX, float mouseY, float customPosX, float customPosY, float partialTicks) {};
+	public void customRender(int mouseX, int mouseY, float customPosX, float customPosY, float partialTicks) {};
 	
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
         return this.isSelected(mouseX, mouseY);
 	}
 	
-	public void hoverCheck(float mouseX, float mouseY) {}
+	public void hoverCheck(int mouseX, int mouseY) {}
 
-    public boolean mouseReleased(double mouseX, double mouseY, int button) {
+    public boolean mouseReleased(int mouseX, int mouseY, int button) {
         return this.isSelected(mouseX, mouseY);
     }
 
     public boolean handleMouseInput() {
-    	float mouseX = Mouse.getEventX() * this.width / MC.displayWidth;
-        float mouseY = this.height - Mouse.getEventY() * this.height / MC.displayHeight - 1;
+    	int mouseX = (int) (Mouse.getEventX() * this.width / MC.displayWidth);
+        int mouseY = (int) (this.height - Mouse.getEventY() * this.height / MC.displayHeight - 1);
         return this.isSelected(mouseX, mouseY);
     }
     
@@ -53,19 +53,19 @@ public abstract class Segment {
     	return false;
     }
     
-    public boolean mouseDragged(double mouseX, double mouseY, int button) {
+    public boolean mouseDragged(int mouseX, int mouseY, int button) {
         return this.isSelected(mouseX, mouseY);
     }
 	
-	public boolean isSelected(double mouseX, double mouseY) {
+	public boolean isSelected(int mouseX, int mouseY) {
         return (((DefaultSettingsGUI) this.gui).popupField == null || this.getIsPopupSegment()) && mouseX >= this.getPosX() && mouseY >= this.getPosY() && mouseX < this.getPosX() + this.getWidth() && mouseY < this.getPosY() + this.getHeight();
 	}
 	
-	public double getPosX() {
+	public float getPosX() {
         return this.posX;
 	}
 	
-	public double getPosY() {
+	public float getPosY() {
         return this.posY;
 	}
 	
@@ -89,7 +89,7 @@ public abstract class Segment {
 	
 	public void guiContentUpdate(String... arg) {};
 	
-	public Segment setPos(double x, double y) {
+	public Segment setPos(float x, float y) {
 		this.posX = x;
 		this.posY = y;
 		return this;
